+++ src_plugins/mincut/pcb-mincut/solve.h
@@ -2,4 +2,4 @@
 
 /* returns a list of object ID pairs (each nth and n+1th element) terminated
    by a -1;-1. Cutting these vertices would separate g. */
-int *solve(gr_t *g, int (*progress)(int so_far, int total, const char *msg));
+int *solve(gr_t *g, int (*progress)(long so_far, long total, const char *msg));
